pybot/pyproject.toml
[project]
name = "dickgrowerbot"
version = "0.1.0"
description = "Python async rewrite of DickGrowerBot"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "aiogram>=3.4,<4",
  "asyncpg>=0.29,<1",
  "python-dotenv>=1.0,<2",
  "PyYAML>=6.0,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.24,<1",
  "ruff>=0.6,<1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
